+++ FlOpEDT/MyFlOp/MyTTModel.py
@@ -43,7 +43,8 @@ class MyTTModel(TTModel):
                  max_stab=5.,
                  lim_ld=1.,
                  core_only=False,
-                 send_mails=False):
+                 send_mails=False,
+                 slots_step=None):
         """
         If you shall change something in the database ahead of creating the
         problem, you must write it here, before calling TTModel's constructor.
@@ -60,7 +61,8 @@ class MyTTModel(TTModel):
                          max_stab=max_stab,
                          lim_ld=lim_ld,
                          core_only=core_only,
-                         send_mails=send_mails)
+                         send_mails=send_mails,
+                         slots_step=slots_step)
 
     def add_specific_constraints(self):
         """