+++ FlOpEDT/TTapp/helpers/minhalfdays.py
@@ -3,21 +3,21 @@
 # This file is part of the FlOpEDT/FlOpScheduler project.
 # Copyright (c) 2017
 # Authors: Iulian Ober, Paul Renaud-Goud, Pablo Seban, et al.
-# 
+#
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Affero General Public License as
 # published by the Free Software Foundation, either version 3 of the
 # License, or (at your option) any later version.
-# 
+#
 # This program is distributed in the hope that it will be useful, but
 # WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
 # Affero General Public License for more details.
-# 
+#
 # You should have received a copy of the GNU Affero General Public
 # License along with this program. If not, see
 # <http://www.gnu.org/licenses/>.
-# 
+#
 # You can be released from the requirements of the license by purchasing
 # a commercial license. Buying such a license is mandatory as soon as
 # you develop activities involving the FlOpEDT/FlOpScheduler software
@@ -26,11 +26,12 @@
 import logging
 
 from base.models import Time, TimeGeneralSettings
-from TTapp.iic.constraint_type import ConstraintType
-from TTapp.iic.constraints.constraint import Constraint
-
+from TTapp.ilp_constraint.constraint_type import ConstraintType
+from TTapp.ilp_constraint.constraint import Constraint
+from TTapp.slots import slots_filter
 
 logger = logging.Logger(__name__)
+
 
 class MinHalfDaysHelperBase():
 
@@ -48,7 +49,7 @@ class MinHalfDaysHelperBase():
     def add_cost(self, cost):
         pass
 
-    
+
     def add_constraint(self, expression, courses, local_var):
         self.ttmodel.add_constraint(local_var, '==', 1,
                                     Constraint(constraint_type=ConstraintType.MIN_HALF_DAYS_LOCAL))
@@ -81,10 +82,10 @@ class MinHalfDaysHelperModule(MinHalfDaysHelperBase):
             mod_b_h_d[(self.module, d, Time.PM)] \
                 = self.ttmodel.add_var("ModBHD(%s,%s,%s)"
                                     % (self.module, d, Time.PM))
-            
+
             # add constraint linking MBHD to TT
             for apm in [Time.AM, Time.PM]:
-                halfdayslots = set(sl for sl in self.ttmodel.wdb.slots if sl.day == d and sl.apm == apm)
+                halfdayslots = set(sl for sl in self.ttmodel.wdb.courses_slots if sl.day == d and sl.apm == apm)
                 card = len(halfdayslots)
                 expr = self.ttmodel.lin_expr()
                 expr += card * mod_b_h_d[(self.module, d, apm)]
@@ -96,7 +97,7 @@ class MinHalfDaysHelperModule(MinHalfDaysHelperBase):
                                             Constraint(constraint_type=ConstraintType.MIN_HALF_DAYS_SUP))
                 self.ttmodel.add_constraint(expr, '<=', card - 1,
                                             Constraint(constraint_type=ConstraintType.MIN_HALF_DAYS_INF))
-        
+
         local_var = self.ttmodel.add_var("MinMBHD_var_%s" % self.module)
         # no year?
         courses = self.ttmodel.wdb.courses.filter(module=self.module, week=self.week)
@@ -108,7 +109,7 @@ class MinHalfDaysHelperModule(MinHalfDaysHelperBase):
         return expression, courses, local_var
 
 
-    def add_cost(self, cost):        
+    def add_cost(self, cost):
         self.ttmodel.obj += cost
 
 
@@ -123,11 +124,12 @@ class MinHalfDaysHelperModule(MinHalfDaysHelperBase):
 class MinHalfDaysHelperGroup(MinHalfDaysHelperBase):
 
     def build_variables(self):
-        courses = self.ttmodel.wdb.courses.filter(group=self.group, week=self.week)
+        courses = self.ttmodel.wdb.courses.filter(groups=self.group, week=self.week)
 
         expression = self.ttmodel.check_and_sum(
             self.ttmodel.GBHD,
-            ((self.group, d, apm) for d, apm in self.ttmodel.wdb.slots_by_half_day if d.week==self.week))
+            ((self.group, d, apm) for apm in self.ttmodel.possible_apms
+             for d in self.ttmodel.wdb.days if d.week == self.week))
 
         local_var = self.ttmodel.add_var("MinGBHD_var_%s" % self.group)
 
@@ -161,7 +163,7 @@ class MinHalfDaysHelperTutor(MinHalfDaysHelperBase):
 
         return expression, courses, local_var
 
-    def add_cost(self, cost):        
+    def add_cost(self, cost):
         self.ttmodel.add_to_inst_cost(self.tutor, cost, self.week)
 
     def add_constraint(self, expression, courses, local_var):
@@ -171,13 +173,13 @@ class MinHalfDaysHelperTutor(MinHalfDaysHelperBase):
         if self.constraint.join2courses and len(courses) in [2, 4]:
             for d in days:
                 for c in courses:
-                    sl8h = min(self.ttmodel.wdb.slots_by_half_day[d,Time.AM] & self.ttmodel.wdb.compatible_slots[c])
-                    sl14h = min(self.ttmodel.wdb.slots_by_half_day[d,Time.PM] & self.ttmodel.wdb.compatible_slots[c])
+                    sl8h = min(slots_filter(self.ttmodel.wdb.courses_slots, day=d, apm=Time.AM) & self.ttmodel.wdb.compatible_slots[c])
+                    sl14h = min(slots_filter(self.ttmodel.wdb.courses_slots, day=d, apm=Time.PM) & self.ttmodel.wdb.compatible_slots[c])
                     for c2 in courses.exclude(id=c.id):
                         sl11h = max(
-                            self.ttmodel.wdb.slots_by_half_day[d, Time.AM] & self.ttmodel.wdb.compatible_slots[c2])
+                            slots_filter(self.ttmodel.wdb.courses_slots, day=d, apm=Time.AM) & self.ttmodel.wdb.compatible_slots[c2])
                         sl17h = max(
-                            self.ttmodel.wdb.slots_by_half_day[d, Time.PM] & self.ttmodel.wdb.compatible_slots[c2])
+                            slots_filter(self.ttmodel.wdb.courses_slots, day=d, apm=Time.PM) & self.ttmodel.wdb.compatible_slots[c2])
                         if self.constraint.weight:
                             conj_var_AM = self.ttmodel.add_conjunct(self.ttmodel.TT[(sl8h, c)],
                                                                     self.ttmodel.TT[(sl11h, c2)])