+++ FlOpEDT/base/models.py
@@ -30,10 +30,8 @@ from django.contrib.auth.models import AbstractUser
 from django.contrib.postgres.fields import ArrayField
 from django.core.validators import MinValueValidator, MaxValueValidator
 from django.db import models
-from django.db.models.signals import pre_save
-from django.dispatch import receiver
 
-from base.timing import hhmm, str_slot
+from base.timing import hhmm, str_slot, Day, Time
 import base.weeks
 
 
@@ -120,62 +118,6 @@ class Group(models.Model):
 # ------------
 # -- TIMING --
 # ------------
-
-class Day(object):
-    MONDAY = "m"
-    TUESDAY = "tu"
-    WEDNESDAY = "w"
-    THURSDAY = "th"
-    FRIDAY = "f"
-    SATURDAY = "sa"
-    SUNDAY = "su"
-
-    CHOICES = ((MONDAY, "monday"), (TUESDAY, "tuesday"),
-               (WEDNESDAY, "wednesday"), (THURSDAY, "thursday"),
-               (FRIDAY, "friday"), (SATURDAY, "saturday"),
-               (SUNDAY, "sunday"))
-
-    def __init__(self, day, week):
-        self.day = day
-        self.week = week
-
-    def __str__(self):
-        # return self.nom[:3]
-        return self.day + '_s' + str(self.week)
-
-
-# will not be used
-# TO BE DELETED at the end
-class Time(models.Model):
-    AM = 'AM'
-    PM = 'PM'
-    HALF_DAY_CHOICES = ((AM, 'AM'), (PM, 'PM'))
-    apm = models.CharField(max_length=2,
-                           choices=HALF_DAY_CHOICES,
-                           verbose_name="Half day",
-                           default=AM)
-    no = models.PositiveSmallIntegerField(default=0)
-    # nom = models.CharField(max_length=20)
-    hours = models.PositiveSmallIntegerField(
-        validators=[MinValueValidator(0), MaxValueValidator(25)], default=8)
-    minutes = models.PositiveSmallIntegerField(
-        validators=[MinValueValidator(0), MaxValueValidator(59)], default=0)
-
-    def __str__(self):
-        return str(self.hours)
-
-    def full_name(self):
-        message = str(self.hours) + ":"
-        if self.minutes < 10:
-            message += "0"
-        message += str(self.minutes)
-        return message
-
-
-@receiver(pre_save, sender=Time)
-def define_apm(sender, instance, *args, **kwargs):
-    if instance.hours >= 12:
-        instance.apm = Time.PM
 
 
 class Holiday(models.Model):
@@ -338,6 +280,19 @@ class ModulePossibleTutors(models.Model):
         'people.Tutor', blank=True, related_name='possible_modules')
 
 
+class ModuleTutorRepartition(models.Model):
+    module = models.ForeignKey('Module', on_delete=models.CASCADE)
+    course_type = models.ForeignKey('CourseType', on_delete=models.CASCADE)
+    tutor = models.ForeignKey('people.Tutor', on_delete=models.CASCADE)
+    week = models.PositiveSmallIntegerField(
+        validators=[MinValueValidator(0), MaxValueValidator(53)],
+        null=True,
+        blank=True)
+    year = models.PositiveSmallIntegerField(null=True,
+                                            blank=True)
+    courses_nb = models.PositiveSmallIntegerField(default=1)
+
+
 class CourseType(models.Model):
     name = models.CharField(max_length=50)
     department = models.ForeignKey(
@@ -364,9 +319,9 @@ class Course(models.Model):
     supp_tutor = models.ManyToManyField('people.Tutor',
                                         related_name='courses_as_supp',
                                         blank=True)
-    group = models.ForeignKey('Group', on_delete=models.CASCADE)
+    groups = models.ManyToManyField('Group', related_name='courses')
     module = models.ForeignKey(
-        'Module', related_name='module', on_delete=models.CASCADE)
+        'Module', related_name='courses', on_delete=models.CASCADE)
     modulesupp = models.ForeignKey('Module', related_name='modulesupp',
                                    null=True, blank=True, on_delete=models.CASCADE)
     week = models.PositiveSmallIntegerField(
@@ -378,18 +333,18 @@ class Course(models.Model):
 
     def __str__(self):
         username_mod = self.tutor.username if self.tutor is not None else '-no_tut-'
-        return f"{self.type}-{self.module}-{username_mod}-{self.group}" \
+        return f"{self.type}-{self.module}-{username_mod}-{'|'.join([g.name for g in self.groups.all()])}" \
                + (" (%s)" % self.id if self.show_id else "")
 
     def full_name(self):
         username_mod = self.tutor.username if self.tutor is not None else '-no_tut-'
-        return f"{self.type}-{self.module}-{username_mod}-{self.group}"
+        return f"{self.type}-{self.module}-{username_mod}-{'|'.join([g.name for g in self.groups.all()])}"
 
     def equals(self, other):
         return self.__class__ == other.__class__ \
                and self.type == other.type \
                and self.tutor == other.tutor \
-               and self.group == other.group \
+               and self.groups == other.groups \
                and self.module == other.module
 
 
@@ -422,6 +377,7 @@ class ScheduledCourse(models.Model):
     def __str__(self):
         return f"{self.course}{self.no}:{self.day}-t{self.start_time}-{self.room}"
 
+    @property
     def end_time(self):
         return self.start_time + self.course.type.duration